requirements.txt
streamlit
langchain
langchain-core
langchain-community
sentence-transformers
faiss-cpu
pypdf
numpy
